move all ethers-utils functions in same file
