--- v-next/hardhat-ethers/src/internal/ethers-utils/types.ts
@@ -1,1 +1,0 @@
-export type FormatFunc = (value: any) => any;
